chore: nettoyage de la racine et archivage des scripts de traduction
