+++ scripts/translations/translate_batch1.py
@@ -1,0 +1,233 @@
+import os
+
+def apply_batch(file_path, translations):
+    if not os.path.exists(file_path): return
+    with open(file_path, 'r', encoding='utf-8') as f: lines = f.readlines()
+    new_lines = []
+    i = 0
+    while i < len(lines):
+        line = lines[i]
+        if line.startswith('msgid "'):
+            msgid = line.split('"')[1]
+            if msgid in translations:
+                new_lines.append(line)
+                i += 1
+                if i < len(lines) and lines[i].startswith('msgstr ""'):
+                    new_lines.append(f'msgstr "{translations[msgid]}"\n')
+                    i += 1
+                    continue
+        new_lines.append(line)
+        i += 1
+    with open(file_path, 'w', encoding='utf-8') as f: f.writelines(new_lines)
+
+# BATCH 1: NL
+batch1_nl = {
+    "<strong>%(count)s</strong> articles publiés": "<strong>%(count)s</strong> artikelen gepubliceerd",
+    "<strong>%(count)s</strong> spectacles trouvés": "<strong>%(count)s</strong> voorstellingen gevonden",
+    "APPROUVÉ": "GOEDGEKEURD",
+    "ATTENTE": "WACHTEND",
+    "Abonnement API Réussi - ThéâtrePlus": "API Abonnement Geslaagd - ThéâtrePlus",
+    "Accepter": "Accepteren",
+    "Accueil": "Home",
+    "Accès API": "API Toegang",
+    "Accès complet aux données (artistes, prix...)": "Volledige toegang tot gegevens (artiesten, prijzen...)",
+    "Accès limité aux données de base": "Beperkte toegang tot basisgegevens",
+    "Action": "Actie",
+    "Actions": "Acties",
+    "Administrateur": "Beheerder",
+    "Administration": "Administratie",
+    "Adresse": "Adres",
+    "Adresse :": "Adres:",
+    "Affecter un nouveau prix": "Nieuwe prijs toewijzen",
+    "Aide": "Help",
+    "Ajouter ce tarif": "Voeg dit tarief toe",
+    "Annuler": "Annuleren",
+    "Annuler et revenir": "Annuleren en terugkeren",
+    "Annuler et revenir au spectacle": "Annuleren en terug naar de voorstelling",
+    "Année de création :": "Jaar van creatie:",
+    "Approuver": "Goedkeuren",
+    "Approuver l'article": "Artikel goedkeuren",
+    "Article de Presse": "Persartikel",
+    "Article introuvable": "Artikel niet gevonden",
+    "Articles de Presse": "Persartikelen",
+    "Articles de presse": "Persartikelen",
+    "Attention :": "Let op:",
+    "Aucun appel enregistré au cours des dernières 24h.": "Geen oproepen geregistreerd in de afgelopen 24 uur.",
+    "Aucun article pour le moment": "Nog geen artikelen",
+    "Aucun avis pour le moment. Soyez le premier à partager votre critique !": "Nog geen beoordelingen. Deel als eerste uw mening!",
+    "Aucun avis à modérer pour vos spectacles.": "Geen beoordelingen om te modereren voor uw voorstellingen.",
+    "Aucun lieu n'est répertorié pour le moment.": "Er zijn momenteel geen locaties geregistreerd.",
+    "Aucun paiement trouvé.": "Geen betalingen gevonden.",
+    "Aucun spectacle disponible pour le moment.": "Geen voorstellingen beschikbaar op dit moment.",
+    "Aucun spectacle en attente.": "Geen voorstellingen in de wacht.",
+    "Aucun spectacle trouvé": "Geen voorstelling gevonden",
+    "Aucun tarif associé.": "Geen tarief gekoppeld.",
+    "Aucun tarif n'est disponible pour ce spectacle.": "Er zijn geen tarieven beschikbaar voor deze voorstelling.",
+    "Aucune candidature en attente.": "Geen kandidaturen in de wacht.",
+    "Aucune demande en attente.": "Geen aanvragen in de wacht.",
+    "Aucune description disponible.": "Geen beschrijving beschikbaar.",
+    "Aucune image fournie.": "Geen afbeelding verstrekt.",
+    "Aucune représentation n'est prévue pour le moment dans ce lieu.": "Er zijn momenteel geen voorstellingen gepland op deze locatie.",
+    "Aucune représentation n'est prévue pour le moment.": "Er zijn momenteel geen voorstellingen gepland.",
+    "Avis épinglé": "Vastgezette beoordeling",
+    "BILLET D'ENTRÉE": "TOEGANGSTICKET",
+    "Besoin de plus de temps ?": "Meer tijd nodig?",
+    "Bienvenue dans le plan": "Welkom bij het plan",
+    "Bienvenue à ThéâtrePlus": "Welkom bij ThéâtrePlus",
+    "Billet": "Ticket",
+    "Billets individuels (QR Codes) :": "Individuele tickets (QR-codes):",
+    "Bonjour,": "Hallo,",
+    "Boostez votre accès": "Verbeter uw toegang",
+    "COMPLET": "VOLGEBOEKT",
+    "Candidat": "Kandidaat",
+    "Capacité de la salle :": "Zaalcapaciteit:",
+    "Catalogue des Spectacles": "Catalogus van voorstellingen",
+    "Ce mois-ci": "Deze maand",
+    "Ce week-end": "Dit weekend",
+    "Cette réservation n'est pas encore payée.": "Deze reservering is nog niet betaald.",
+    "Cette semaine": "Deze week",
+    "Choisir ce plan": "Kies dit plan",
+    "Choisir un type de prix": "Kies een prijstype",
+    "Choisir une salle": "Kies een zaal",
+    "Client": "Klant",
+    "Clé API copiée dans le presse-papier !": "API-sleutel gekopieerd naar klembord!",
+    "Commande annulée": "Bestelling geannuleerd",
+    "Commande annulée - ThéâtrePlus": "Bestelling geannuleerd - ThéâtrePlus",
+    "Commande validée !": "Bestelling bevestigd!",
+    "Commencer maintenant": "Nu beginnen",
+    "Complet": "Volgeboekt",
+    "Compte": "Account",
+    "Comédie": "Komedie",
+    "Concert": "Concert",
+    "Concerts": "Concerten",
+    "Concerts, théâtre, spectacles et bien plus encore": "Concerten, theater, shows en nog veel meer",
+    "Conditions": "Voorwaarden",
+    "Confidentialité": "Privacy",
+    "Confirmation de commande": "Bestelbevestiging",
+    "Confirmation de votre commande": "Bevestiging van uw bestelling",
+    "Confirmer et Payer": "Bevestigen en betalen",
+    "Confirmer la suppression": "Verwijdering bevestigen",
+    "Connexion": "Inloggen",
+    "Consommation API - ThéâtrePlus": "API Verbruik - ThéâtrePlus",
+    "Consommation de l'API": "API Verbruik",
+    "Consultez et validez les retours de vos spectateurs.": "Bekijk en valideer de feedback van uw toeschouwers.",
+    "Contact": "Contact",
+    "Continuer mes achats": "Verder winkelen",
+    "Copier": "Kopiëren",
+    "Critique de Presse": "Persrecensent",
+    "Critique inexistante": "Recensie bestaat niet",
+    "Critique invité": "Gastrecensent",
+    "Critique modifiée avec succès.": "Recensie succesvol gewijzigd.",
+    "Critiques & Avis": "Recensies & Beoordelingen",
+    "Créer un nouveau type de prix": "Nieuw prijstype aanmaken",
+    "DERNIÈRES PLACES :": "LAATSTE PLAATSEN:",
+    "Danse": "Dans",
+}
+
+# BATCH 1: EN
+batch1_en = {
+    "<strong>%(count)s</strong> articles publiés": "<strong>%(count)s</strong> articles published",
+    "<strong>%(count)s</strong> spectacles trouvés": "<strong>%(count)s</strong> shows found",
+    "APPROUVÉ": "APPROVED",
+    "ATTENTE": "PENDING",
+    "Abonnement API Réussi - ThéâtrePlus": "API Subscription Successful - ThéâtrePlus",
+    "Accepter": "Accept",
+    "Accueil": "Home",
+    "Accès API": "API Access",
+    "Accès complet aux données (artistes, prix...)": "Full access to data (artists, prices...)",
+    "Accès limitété aux données de base": "Limited access to basic data",
+    "Action": "Action",
+    "Actions": "Actions",
+    "Administrateur": "Administrator",
+    "Administration": "Administration",
+    "Adresse": "Address",
+    "Adresse :": "Address:",
+    "Affecter un nouveau prix": "Assign a new price",
+    "Aide": "Help",
+    "Ajouter ce tarif": "Add this rate",
+    "Annuler": "Cancel",
+    "Annuler et revenir": "Cancel and return",
+    "Annuler et revenir au spectacle": "Cancel and return to show",
+    "Année de création :": "Creation year:",
+    "Approuver": "Approve",
+    "Approuver l'article": "Approve article",
+    "Article de Presse": "Press Article",
+    "Article introuvable": "Article not found",
+    "Articles de Presse": "Press Articles",
+    "Articles de presse": "Press Articles",
+    "Attention :": "Attention:",
+    "Aucun appel enregistré au cours des dernières 24h.": "No calls recorded in the last 24 hours.",
+    "Aucun article pour le moment": "No articles yet",
+    "Aucun avis pour le moment. Soyez le premier à partager votre critique !": "No reviews yet. Be the first to share yours!",
+    "Aucun avis à modérer pour vos spectacles.": "No reviews to moderate for your shows.",
+    "Aucun lieu n'est répertorié pour le moment.": "No locations listed yet.",
+    "Aucun paiement trouvé.": "No payments found.",
+    "Aucun spectacle disponible pour le moment.": "No shows available yet.",
+    "Aucun spectacle en attente.": "No pending shows.",
+    "Aucun spectacle trouvé": "No show found",
+    "Aucun tarif associé.": "No rate associated.",
+    "Aucun tarif n'est disponible pour ce spectacle.": "No rates available for this show.",
+    "Aucune candidature en attente.": "No pending applications.",
+    "Aucune demande en attente.": "No pending requests.",
+    "Aucune description disponible.": "No description available.",
+    "Aucune image fournie.": "No image provided.",
+    "Aucune représentation n'est prévue pour le moment dans ce lieu.": "No representations scheduled yet at this location.",
+    "Aucune représentation n'est prévue pour le moment.": "No representations scheduled yet.",
+    "Avis épinglé": "Pinned review",
+    "BILLET D'ENTRÉE": "ENTRY TICKET",
+    "Besoin de plus de temps ?": "Need more time?",
+    "Bienvenue dans le plan": "Welcome to the plan",
+    "Bienvenue à ThéâtrePlus": "Welcome to ThéâtrePlus",
+    "Billet": "Ticket",
+    "Billets individuels (QR Codes) :": "Individual tickets (QR Codes):",
+    "Bonjour,": "Hello,",
+    "Boostez votre accès": "Boost your access",
+    "COMPLET": "SOLD OUT",
+    "Candidat": "Candidate",
+    "Capacité de la salle :": "Room capacity:",
+    "Catalogue des Spectacles": "Show Catalogue",
+    "Ce mois-ci": "This month",
+    "Ce week-end": "This weekend",
+    "Cette réservation n'est pas encore payée.": "This reservation is not paid yet.",
+    "Cette semaine": "This week",
+    "Choisir ce plan": "Choose this plan",
+    "Choisir un type de prix": "Choose a price type",
+    "Choisir une salle": "Choose a room",
+    "Client": "Customer",
+    "Clé API copiée dans le presse-papier !": "API Key copied to clipboard!",
+    "Commande annulée": "Order cancelled",
+    "Commande annulée - ThéâtrePlus": "Order cancelled - ThéâtrePlus",
+    "Commande validée !": "Order confirmed!",
+    "Commencer maintenant": "Start now",
+    "Complet": "Sold out",
+    "Compte": "Account",
+    "Comédie": "Comedy",
+    "Concert": "Concert",
+    "Concerts": "Concerts",
+    "Concerts, théâtre, spectacles et bien plus encore": "Concerts, theater, shows and much more",
+    "Conditions": "Terms",
+    "Confidentialité": "Privacy",
+    "Confirmation de commande": "Order confirmation",
+    "Confirmation de votre commande": "Confirmation of your order",
+    "Confirmer et Payer": "Confirm and Pay",
+    "Confirmer la suppression": "Confirm deletion",
+    "Connexion": "Login",
+    "Consommation API - ThéâtrePlus": "API Usage - ThéâtrePlus",
+    "Consommation de l'API": "API Usage",
+    "Consultez et validez les retours de vos spectateurs.": "Check and validate your audience feedback.",
+    "Contact": "Contact",
+    "Continuer mes achats": "Continue shopping",
+    "Copier": "Copy",
+    "Critique de Presse": "Press Critic",
+    "Critique inexistante": "Review does not exist",
+    "Critique invité": "Guest critic",
+    "Critique modifiée avec succès.": "Review updated successfully.",
+    "Critiques & Avis": "Reviews & Feedback",
+    "Créer un nouveau type de prix": "Create a new price type",
+    "DERNIÈRES PLACES :": "LAST SEATS:",
+    "Danse": "Dance",
+}
+
+apply_batch('locale/nl/LC_MESSAGES/django.po', batch1_nl)
+apply_batch('locale/en/LC_MESSAGES/django.po', batch1_en)
+print("Batch 1 (A-D) applied.")
