+++ scripts/translations/translate_batch3.py
@@ -1,0 +1,214 @@
+import os
+from translate_batch1 import apply_batch
+
+# BATCH 3: NL
+batch3_nl = {
+    "Maximum 500 caractères.": "Maximaal 500 tekens.",
+    "Membre": "Lid",
+    "Merci pour votre réservation": "Bedankt voor uw reservering",
+    "Mes Articles de Presse": "Mijn persartikelen",
+    "Mes Réservations": "Mijn reserveringen",
+    "Mes Spectacles": "Mijn voorstellingen",
+    "Mes articles": "Mijn artikelen",
+    "Mes articles de presse": "Mijn persartikelen",
+    "Message supprimé avec succès.": "Bericht succesvol verwijderd.",
+    "Mettez à jour les informations de votre spectacle.": "Werk de informatie van uw voorstelling bij.",
+    "Mise à jour pour": "Update voor",
+    "Modification de critique": "Wijziging van recensie",
+    "Modifier": "Bewerken",
+    "Modifier dans le Dashboard Admin": "Bewerken in het Admin Dashboard",
+    "Modifier dans mon Espace Producteur": "Bewerken in mijn Producentenruimte",
+    "Modifier le profil": "Profiel bewerken",
+    "Modifier le spectacle": "Voorstelling bewerken",
+    "Modifier ma critique": "Mijn recensie bewerken",
+    "Modifier mon avis": "Mijn beoordeling bewerken",
+    "Modifier mon panier": "Mijn winkelmandje bewerken",
+    "Modération des avis": "Beoordelingen modereren",
+    "Modérer les avis": "Beoordelingen modereren",
+    "Modérez les articles rédigés par les critiques sur vos spectacles.": "Modereer de artikelen geschreven door recensenten over uw voorstellingen.",
+    "Mon Accès API - ThéâtrePlus": "Mijn API-toegang - ThéâtrePlus",
+    "Mon Billet - ThéâtrePlus": "Mijn ticket - ThéâtrePlus",
+    "Mon Panier": "Mijn Winkelmandje",
+    "Mon Profil": "Mijn Profiel",
+    "Mon Profil - ThéâtrePlus": "Mijn Profiel - ThéâtrePlus",
+    "Mon compte": "Mijn account",
+    "Mon profil": "Mijn profiel",
+    "Montant": "Bedrag",
+    "Montant Total :": "Totaalbedrag:",
+    "Montant à débiter": "Te debiteren bedrag",
+    "Mot de passe": "Wachtwoord",
+    "Motivation": "Motivatie",
+    "Média / Journal": "Media / Krant",
+    "Méthode": "Methode",
+    "Navigation": "Navigatie",
+    "Nederlands": "Nederlands",
+    "Nom": "Naam",
+    "Nom complet": "Volledige naam",
+    "Nombre de tickets": "Aantal tickets",
+    "Nombre de tickets autorisés": "Toegestaan aantal tickets",
+    "Non catégorisé": "Niet gecategoriseerd",
+    "Non défini": "Niet gedefinieerd",
+    "Non définie": "Niet gedefinieerd",
+    "Non fourni": "Niet verstrekt",
+    "Non réservable": "Niet boekbaar",
+    "Non, garder mon avis": "Nee, bewaar mijn beoordeling",
+    "Nos Lieux": "Onze Locaties",
+    "Nos Lieux de Spectacles": "Onze Theaterlocaties",
+    "Nos Spectacles": "Onze Voorstellingen",
+    "Note": "Cijfer",
+    "Note du Producteur": "Producentennote",
+    "Notre Histoire": "Onze Geschiedenis",
+    "Notre Mission": "Onze Missie",
+    "OFFERT": "GRATIS",
+    "Oui, supprimer définitivement": "Ja, definitief verwijderen",
+    "Ouvrir Ticketmaster": "Open Ticketmaster",
+    "PAYÉ": "BETAALD",
+    "PUBLIÉ": "GEPUBLICEERD",
+    "Page": "Pagina",
+    "Paiement Réussi !": "Betaling Geslaagd!",
+    "Paiement réussi - ThéâtrePlus": "Betaling geslaagd - ThéâtrePlus",
+    "Paiement sécurisé via notre plateforme": "Veilig betalen via ons platform",
+    "Par": "Door",
+    "Partager un spectacle": "Een voorstelling delen",
+    "Partagez vos analyses et critiques professionnelles sur ThéâtrePlus.": "Deel uw professionele analyses en recensies op ThéâtrePlus.",
+    "Partagez votre avis sur": "Deel uw mening over",
+    "Passerelle de Paiement": "Betalingsgateway",
+    "Places": "Plaatsen",
+    "Places restantes": "Resterende plaatsen",
+    "Plan Actuel": "Huidig Plan",
+    "Plan actif": "Actief plan",
+    "Plan actuel :": "Huidig plan:",
+    "Première Séance": "Eerste sessie",
+    "Prix": "Prijs",
+    "Prix :": "Prijs:",
+    "Prochain spectacle": "Volgende voorstelling",
+    "Prochaines représentations": "Volgende voorstellingen",
+    "Producteur": "Producent",
+    "Profession": "Beroep",
+    "Proposez vos propres spectacles sur ThéâtrePlus.": "Stel uw eigen voorstellingen voor op ThéâtrePlus.",
+    "Proposez votre spectacle à l'administration.": "Stel uw voorstelling voor aan de administratie.",
+    "Précédent": "Vorige",
+    "Prénom": "Voornaam",
+    "Présentation": "Presentatie",
+    "Présentez-vous": "Stel jezelf voor",
+    "Présentez-vous brièvement (max 500 caractères)...": "Stel jezelf kort voor (max 500 tekens)...",
+    "Publier l'article": "Artikel publiceren",
+    "Publier le Spectacle": "Voorstelling publiceren",
+    "Publier ma critique": "Mijn recensie publiceren",
+    "Publier maintenant": "Nu publiceren",
+    "Publié": "Gepubliceerd",
+    "Qu'avez-vous l'intention de proposer ?": "Wat bent u van plan voor te stellen?",
+    "Quantité": "Aantal",
+    "Quelle note donneriez-vous ?": "Welk cijfer zou u geven?",
+    "REJETÉ": "AFGEWEZEN",
+    "Rechercher": "Zoeken",
+    "Rechercher par utilisateur ou ID Stripe...": "Zoeken op gebruiker of Stripe ID...",
+}
+
+# BATCH 3: EN
+batch3_en = {
+    "Maximum 500 caractères.": "Maximum 500 characters.",
+    "Membre": "Member",
+    "Merci pour votre réservation": "Thank you for your booking",
+    "Mes Articles de Presse": "My Press Articles",
+    "Mes Réservations": "My Bookings",
+    "Mes Spectacles": "My Shows",
+    "Mes articles": "My articles",
+    "Mes articles de presse": "My press articles",
+    "Message supprimé avec succès.": "Message deleted successfully.",
+    "Mettez à jour les informations de votre spectacle.": "Update your show information.",
+    "Mise à jour pour": "Update for",
+    "Modification de critique": "Review edit",
+    "Modifier": "Edit",
+    "Modifier dans le Dashboard Admin": "Edit in Admin Dashboard",
+    "Modifier dans mon Espace Producteur": "Edit in my Producer Space",
+    "Modifier le profil": "Edit profile",
+    "Modifier le spectacle": "Edit show",
+    "Modifier ma critique": "Edit my review",
+    "Modifier mon avis": "Edit my review",
+    "Modifier mon panier": "Edit my cart",
+    "Modération des avis": "Review moderation",
+    "Modérer les avis": "Moderate reviews",
+    "Modérez les articles rédigés par les critiques sur vos spectacles.": "Moderate articles written by critics about your shows.",
+    "Mon Accès API - ThéâtrePlus": "My API Access - ThéâtrePlus",
+    "Mon Billet - ThéâtrePlus": "My Ticket - ThéâtrePlus",
+    "Mon Panier": "My Cart",
+    "Mon Profil": "My Profile",
+    "Mon Profil - ThéâtrePlus": "My Profile - ThéâtrePlus",
+    "Mon compte": "My account",
+    "Mon profil": "My profile",
+    "Montant": "Amount",
+    "Montant Total :": "Total Amount:",
+    "Montant à débiter": "Amount to charge",
+    "Mot de passe": "Password",
+    "Motivation": "Motivation",
+    "Média / Journal": "Media / Journal",
+    "Méthode": "Method",
+    "Navigation": "Navigation",
+    "Nederlands": "Dutch",
+    "Nom": "Last name",
+    "Nom complet": "Full name",
+    "Nombre de tickets": "Number of tickets",
+    "Nombre de tickets autorisés": "Allowed tickets",
+    "Non catégorisé": "Uncategorized",
+    "Non défini": "Not defined",
+    "Non définie": "Not defined",
+    "Non fourni": "Not provided",
+    "Non réservable": "Not bookable",
+    "Non, garder mon avis": "No, keep my review",
+    "Nos Lieux": "Our Venues",
+    "Nos Lieux de Spectacles": "Our Performance Venues",
+    "Nos Spectacles": "Our Shows",
+    "Note": "Rating",
+    "Note du Producteur": "Producer Note",
+    "Notre Histoire": "Our History",
+    "Notre Mission": "Our Mission",
+    "OFFERT": "FREE",
+    "Oui, supprimer définitivement": "Yes, permanently delete",
+    "Ouvrir Ticketmaster": "Open Ticketmaster",
+    "PAYÉ": "PAID",
+    "PUBLIÉ": "PUBLISHED",
+    "Page": "Page",
+    "Paiement Réussi !": "Payment Successful!",
+    "Paiement réussi - ThéâtrePlus": "Payment successful - ThéâtrePlus",
+    "Paiement sécurisé via notre plateforme": "Secure payment via our platform",
+    "Par": "By",
+    "Partager un spectacle": "Share a show",
+    "Partagez vos analyses et critiques professionnelles sur ThéâtrePlus.": "Share your professional analyses and reviews on ThéâtrePlus.",
+    "Partagez votre avis sur": "Share your review on",
+    "Passerelle de Paiement": "Payment Gateway",
+    "Places": "Seats",
+    "Places restantes": "Remaining seats",
+    "Plan Actuel": "Current Plan",
+    "Plan actif": "Active plan",
+    "Plan actuel :": "Current plan:",
+    "Première Séance": "First session",
+    "Prix": "Price",
+    "Prix :": "Price:",
+    "Prochain spectacle": "Next show",
+    "Prochaines représentations": "Next representations",
+    "Producteur": "Producer",
+    "Profession": "Profession",
+    "Proposez vos propres spectacles sur ThéâtrePlus.": "Offer your own shows on ThéâtrePlus.",
+    "Proposez votre spectacle à l'administration.": "Submit your show to administration.",
+    "Précédent": "Previous",
+    "Prénom": "First name",
+    "Présentation": "Presentation",
+    "Présentez-vous": "Introduce yourself",
+    "Présentez-vous brièvement (max 500 caractères)...": "Introduce yourself briefly (max 500 chars)...",
+    "Publier l'article": "Publish article",
+    "Publier le Spectacle": "Publish show",
+    "Publier ma critique": "Publish my review",
+    "Publier maintenant": "Publish now",
+    "Publié": "Published",
+    "Qu'avez-vous l'intention de proposer ?": "What do you intend to offer?",
+    "Quantité": "Quantity",
+    "Quelle note donneriez-vous ?": "What rating would you give?",
+    "REJETÉ": "REJECTED",
+    "Rechercher": "Search",
+    "Rechercher par utilisateur ou ID Stripe...": "Search by user or Stripe ID...",
+}
+
+apply_batch('locale/nl/LC_MESSAGES/django.po', batch3_nl)
+apply_batch('locale/en/LC_MESSAGES/django.po', batch3_en)
+print("Batch 3 (M-R) applied.")
