+++ scripts/translations/translate_batch4.py
@@ -1,0 +1,334 @@
+import os
+from translate_batch1 import apply_batch
+
+# BATCH 4: NL
+batch4_nl = {
+    "Rechercher un spectacle, artiste ou lieu...": "Zoek een voorstelling, artiest of locatie...",
+    "Refuser": "Weigeren",
+    "Rejeter": "Afwijzen",
+    "Retirer de la publication": "Uit publicatie halen",
+    "Retour au spectacle": "Terug naar voorstelling",
+    "Retour au tableau de bord": "Terug naar dashboard",
+    "Retour aux articles": "Terug naar artikelen",
+    "Retour à l'accueil": "Terug naar home",
+    "Retour à la liste": "Terug naar lijst",
+    "Retour à la liste des lieux": "Terug naar locatielijst",
+    "Retour à mes réservations": "Terug naar mijn reserveringen",
+    "Retour à mon profil": "Terug naar mijn profiel",
+    "Retourner au panier": "Terug naar winkelmandje",
+    "Retourner à l'Espace API": "Terug naar API-ruimte",
+    "Revenez bientôt pour découvrir nos dernières critiques de presse.": "Kom snel terug om onze nieuwste persrecensies te ontdekken.",
+    "Récapitulatif": "Overzicht",
+    "Récapitulatif des billets": "Ticketoverzicht",
+    "Rédiger un article": "Een artikel schrijven",
+    "Réessayer le paiement": "Betaling opnieuw proberen",
+    "Réponse": "Antwoord",
+    "Réservable": "Boekbaar",
+    "Réservation": "Reservering",
+    "Réserver": "Reserveren",
+    "Réserver mes places": "Mijn plaatsen reserveren",
+    "Réservez vos billets dès maintenant": "Reserveer nu uw tickets",
+    "Réservez vos billets pour les meilleurs événements": "Reserveer uw tickets voor de beste evenementen",
+    "Réservez vos spectacles": "Reserveer uw voorstellingen",
+    "Résumé": "Samenvatting",
+    "Résumé de vos avantages :": "Overzicht van uw voordelen:",
+    "S'abonner": "Abonneren",
+    "Salle / Lieu": "Zaal / Locatie",
+    "Seul l'auteur peut modifier son message.": "Alleen de auteur kan dit bericht wijzigen.",
+    "Simulation de la transaction pour la réservation": "Simulatie van de transactie voor de reservering",
+    "Simulation de paiement": "Betalingssimulatie",
+    "Simuler un PAIEMENT RÉUSSI": "Simuleer een GESLAAGDE BETALING",
+    "Simuler un ÉCHEC DE PAIEMENT": "Simuleer een MISLUKTE BETALING",
+    "Site Web": "Website",
+    "Site Web :": "Website:",
+    "Soumettre ma demande": "Mijn aanvraag indienen",
+    "Soumettre une critique": "Een recensie indienen",
+    "Soumettre votre candidature": "Uw kandidatuur indienen",
+    "Soumettre votre demande": "Uw aanvraag indienen",
+    "Soumission de critique": "Recensie indienen",
+    "Sous-total": "Subtotaal",
+    "Spectacle": "Voorstelling",
+    "Spectacle / Séance": "Voorstelling / Sessie",
+    "Spectacle créé avec succès.": "Voorstelling succesvol aangemaakt.",
+    "Spectacle inexistant": "Voorstelling bestaat niet",
+    "Spectacle introuvable": "Voorstelling niet gevonden",
+    "Spectacle modifié avec succès.": "Voorstelling succesvol gewijzigd.",
+    "Spectacle supprimé avec succès.": "Voorstelling succesvol verwijderd.",
+    "Spectacle(s)": "Voorstelling(en)",
+    "Spectacles": "Voorstellingen",
+    "Statistiques d'utilisation": "Gebruiksstatistieken",
+    "Statut": "Status",
+    "Succès": "Succes",
+    "Suivant": "Volgende",
+    "Suivez l'utilisation de vos quotas en temps réel.": "Volg uw quotaverbruik in realtime.",
+    "Support prioritaire": "Prioritaire ondersteuning",
+    "Support technique prioritaire": "Prioritaire technische ondersteuning",
+    "Suppression d'un autre compte interdite!": "Verwijderen van een ander account niet toegestaan!",
+    "Suppression interdite (méthode incorrecte)!": "Verwijderen niet toegestaan (onjuiste methode)!",
+    "Supprimer": "Verwijderen",
+    "Supprimer cet avis ?": "Deze beoordeling verwijderen?",
+    "Supprimer définitivement cet article ?": "Dit artikel definitief verwijderen?",
+    "Supprimer mon compte": "Mijn account verwijderen",
+    "Sélectionnez le spectacle sur lequel porte votre article.": "Selecteer de voorstelling waarover uw artikel gaat.",
+    "Sélectionnez une photo depuis votre ordinateur.": "Selecteer een foto van uw computer.",
+    "Sélectionnez vos places": "Selecteer uw plaatsen",
+    "TOTAL À PAYER :": "TOTAAL TE BETALEN:",
+    "Tarif": "Tarief",
+    "Théâtre": "Theater",
+    "Tickets vendus": "Verkochte tickets",
+    "Titre": "Titel",
+    "Titre du spectacle": "Titel van de voorstelling",
+    "Total": "Totaal",
+    "Tous droits réservés. | Projet PID - Version Alpha": "Alle rechten voorbehouden. | Project PID - Alpha-versie",
+    "Tous les genres": "Alle genres",
+    "Tous les événements": "Alle evenementen",
+    "Toutes les dates": "Alle data",
+    "Type de place": "Type plaats",
+    "Télécharger mon billet (PDF)": "Download mijn ticket (PDF)",
+    "Télécharger tous les billets (PDF)": "Download alle tickets (PDF)",
+    "Téléphone": "Telefoon",
+    "Téléphone :": "Telefoon:",
+    "Un court texte d'accroche qui sera visible dans les listes.": "Een korte introductietekst die zichtbaar zal zijn in lijsten.",
+    "Un e-mail récapitulatif vous a également été envoyé.": "Er is ook een e-mail met een overzicht naar u verzonden.",
+    "Un exemplaire de vos billets a été envoyé à votre adresse e-mail": "Een kopie van uw tickets is verzonden naar uw e-mailadres",
+    "Un problème est survenu": "Er is een probleem opgetreden",
+    "Un spectacle doit avoir au moins un tarif pour être publié.": "Een voorstelling moet minstens één tarief hebben om te worden gepubliceerd.",
+    "Utilisateur": "Gebruiker",
+    "Utilisateur supprimé avec succès.": "Gebruiker succesvol verwijderd.",
+    "Utilisation aujourd'hui": "Gebruik vandaag",
+    "VOIR LE PANIER": "WINKELMANDJE BEKIJKEN",
+    "Valider la commande": "Bestelling bevestigen",
+    "Veuillez choisir un spectacle pour laisser un commentaire.": "Kies een voorstelling om een reactie achter te laten.",
+    "Veuillez choisir une note.": "Kies een cijfer.",
+    "Veuillez corriger les erreurs ci-dessous.": "Corrigeer de onderstaande fouten.",
+    "Veuillez vérifier vos informations bancaires et réessayer.": "Controleer uw bankgegevens en probeer het opnieuw.",
+    "Visiter le site officiel": "Bezoek de officiële site",
+    "Voir Tout": "Alles bekijken",
+    "Voir le spectacle": "Voorstelling bekijken",
+    "Voir mes billets": "Mijn tickets bekijken",
+    "Voir mon panier": "Mijn winkelmandje bekijken",
+    "Voir sur le site": "Op de site bekijken",
+    "Vos billets officiels - ThéâtrePlus": "Uw officiële tickets - ThéâtrePlus",
+    "Votre Clé API": "Uw API-sleutel",
+    "Votre Critique": "Uw Recensie",
+    "Votre Panier": "Uw Winkelmandje",
+    "Votre Panier - ThéâtrePlus": "Uw Winkelmandje - ThéâtrePlus",
+    "Votre abonnement API a été mis à jour avec succès.": "Uw API-abonnement is succesvol bijgewerkt.",
+    "Votre avis a été soumis avec succès et sera publié après validation.": "Uw beoordeling is succesvol ingediend en zal na validatie worden gepubliceerd.",
+    "Votre commentaire": "Uw reactie",
+    "Votre motivation": "Uw motivatie",
+    "Votre paiement a été traité avec succès.": "Uw betaling is succesvol verwerkt.",
+    "Votre panier est vide": "Uw winkelmandje is leeg",
+    "Votre panier est vide.": "Uw winkelmandje is leeg.",
+    "Votre quota est réinitialisé chaque jour à minuit.": "Uw quota wordt elke dag om middernacht gereset.",
+    "Votre transaction a été refusée par l'organisme bancaire.": "Uw transactie is geweigerd door de bank.",
+    "Vous avez choisi d'interrompre le processus de paiement.": "U heeft ervoor gekozen om het betalingsproces te onderbreken.",
+    "Vous n'avez pas encore activé votre accès API.": "U heeft uw API-toegang nog niet geactiveerd.",
+    "Vous n'avez pas encore effectué de réservation.": "U heeft nog geen reservering gemaakt.",
+    "Vous n'avez pas encore partagé de spectacle.": "U heeft nog geen voorstelling gedeeld.",
+    "Vous n'avez pas encore rédigé d'article.": "U heeft nog geen artikel geschreven.",
+    "Vous n'avez pas l'autorisation d'accéder à cette page!": "U heeft geen toestemming om deze pagina te bekijken!",
+    "Vous n'êtes pas autorisé à supprimer ce message.": "U bent niet gemachtigd om dit bericht te verwijderen.",
+    "Vous ne pouvez commenter que les spectacles auxquels vous avez assisté.": "U kunt alleen reageren op voorstellingen die u heeft bezocht.",
+    "Vous pouvez les consulter même si vous n'avez plus accès à internet.": "U kunt ze bekijken, zelfs als u geen internettoegang meer heeft.",
+    "Vous êtes déjà critique de presse.": "U bent al persrecensent.",
+    "Vous êtes déjà inscrit!": "U bent al geregistreerd!",
+    "Vous êtes déjà producteur.": "U bent al producent.",
+    "appels / jour": "oproepen / dag",
+    "appels par jour.": "oproepen per dag.",
+    "du": "van",
+    "minutes": "minuten",
+    "mois": "maand",
+    "place(s) à": "plaats(en) aan",
+    "places": "plaatsen",
+    "places disponibles": "beschikbare plaatsen",
+    "sur": "op",
+    "À propos": "Over ons",
+    "À propos de ThéâtrePlus": "Over ThéâtrePlus",
+    "ÉCHEC": "MISLUKT",
+    "ÉCHOUÉ": "MISLUKT",
+    "ÉPINGLÉ": "VASTGEZET",
+    "Échec du Paiement": "Betaling Mislukt",
+    "Échec du paiement - ThéâtrePlus": "Betaling mislukt - ThéâtrePlus",
+    "Éditer": "Bewerken",
+    "Éditer (Admin)": "Bewerken (Admin)",
+    "Épingler": "Vastzetten",
+    "Épingler en haut": "Bovenaan vastzetten",
+    "Épinglé": "Vastgezet",
+    "État": "Status",
+    "Événement": "Evenement",
+    "Événements à venir": "Komende evenementen",
+    "Êtes-vous sûr ?": "Weet u het zeker?",
+    "Êtes-vous sûr de vouloir refuser cette candidature ?": "Weet u zeker dat u deze kandidatuur wilt weigeren?",
+    "Êtes-vous sûr de vouloir refuser cette demande ?": "Weet u zeker dat u deze aanvraag wilt weigeren?",
+    "à": "te",
+    "à partir de": "vanaf",
+}
+
+# BATCH 4: EN
+batch4_en = {
+    "Rechercher un spectacle, artiste ou lieu...": "Search for a show, artist or venue...",
+    "Refuser": "Reject",
+    "Rejeter": "Reject",
+    "Retirer de la publication": "Remove from publication",
+    "Retour au spectacle": "Back to show",
+    "Retour au tableau de bord": "Back to dashboard",
+    "Retour aux articles": "Back to articles",
+    "Retour à l'accueil": "Back to home",
+    "Retour à la liste": "Back to list",
+    "Retour à la liste des lieux": "Back to venues list",
+    "Retour à mes réservations": "Back to my bookings",
+    "Retour à mon profil": "Back to my profile",
+    "Retourner au panier": "Back to cart",
+    "Retourner à l'Espace API": "Return to API Space",
+    "Revenez bientôt pour découvrir nos dernières critiques de presse.": "Come back soon to discover our latest press reviews.",
+    "Récapitulatif": "Summary",
+    "Récapitulatif des billets": "Ticket summary",
+    "Rédiger un article": "Write an article",
+    "Réessayer le paiement": "Retry payment",
+    "Réponse": "Response",
+    "Réservable": "Bookable",
+    "Réservation": "Booking",
+    "Réserver": "Book",
+    "Réserver mes places": "Book my seats",
+    "Réservez vos billets dès maintenant": "Book your tickets now",
+    "Réservez vos billets pour les meilleurs événements": "Book your tickets for the best events",
+    "Réservez vos spectacles": "Book your shows",
+    "Résumé": "Summary",
+    "Résumé de vos avantages :": "Summary of your benefits:",
+    "S'abonner": "Subscribe",
+    "Salle / Lieu": "Hall / Venue",
+    "Seul l'auteur peut modifier son message.": "Only the author can edit this message.",
+    "Simulation de la transaction pour la réservation": "Transaction simulation for the booking",
+    "Simulation de paiement": "Payment simulation",
+    "Simuler un PAIEMENT RÉUSSI": "Simulate SUCCESSFUL PAYMENT",
+    "Simuler un ÉCHEC DE PAIEMENT": "Simulate PAYMENT FAILURE",
+    "Site Web": "Website",
+    "Site Web :": "Website:",
+    "Soumettre ma demande": "Submit my request",
+    "Soumettre une critique": "Submit a review",
+    "Soumettre votre candidature": "Submit your application",
+    "Soumettre votre demande": "Submit your request",
+    "Soumission de critique": "Review submission",
+    "Sous-total": "Subtotal",
+    "Spectacle": "Show",
+    "Spectacle / Séance": "Show / Session",
+    "Spectacle créé avec succès.": "Show created successfully.",
+    "Spectacle inexistant": "Show does not exist",
+    "Spectacle introuvable": "Show not found",
+    "Spectacle modifié avec succès.": "Show updated successfully.",
+    "Spectacle supprimé avec succès.": "Show deleted successfully.",
+    "Spectacle(s)": "Show(s)",
+    "Spectacles": "Shows",
+    "Statistiques d'utilisation": "Usage statistics",
+    "Statut": "Status",
+    "Succès": "Success",
+    "Suivant": "Next",
+    "Suivez l'utilisation de vos quotas en temps réel.": "Track your quota usage in real time.",
+    "Support prioritaire": "Priority support",
+    "Support technique prioritaire": "Priority technical support",
+    "Suppression d'un autre compte interdite!": "Deletion of another account forbidden!",
+    "Suppression interdite (méthode incorrecte)!": "Deletion forbidden (incorrect method)!",
+    "Supprimer": "Delete",
+    "Supprimer cet avis ?": "Delete this review?",
+    "Supprimer définitivement cet article ?": "Permanently delete this article?",
+    "Supprimer mon compte": "Delete my account",
+    "Sélectionnez le spectacle sur lequel porte votre article.": "Select the show your article is about.",
+    "Sélectionnez une photo depuis votre ordinateur.": "Select a photo from your computer.",
+    "Sélectionnez vos places": "Select your seats",
+    "TOTAL À PAYER :": "TOTAL TO PAY:",
+    "Tarif": "Rate",
+    "Théâtre": "Theater",
+    "Tickets vendus": "Sold tickets",
+    "Titre": "Title",
+    "Titre du spectacle": "Show title",
+    "Total": "Total",
+    "Tous droits réservés. | Projet PID - Version Alpha": "All rights reserved. | PID Project - Alpha Version",
+    "Tous les genres": "All genres",
+    "Tous les événements": "All events",
+    "Toutes les dates": "All dates",
+    "Type de place": "Seat type",
+    "Télécharger mon billet (PDF)": "Download my ticket (PDF)",
+    "Télécharger tous les billets (PDF)": "Download all tickets (PDF)",
+    "Téléphone": "Phone",
+    "Téléphone :": "Phone:",
+    "Un court texte d'accroche qui sera visible dans les listes.": "A short intro text that will be visible in lists.",
+    "Un e-mail récapitulatif vous a également été envoyé.": "A summary email has also been sent to you.",
+    "Un exemplaire de vos billets a été envoyé à votre adresse e-mail": "A copy of your tickets has been sent to your email address",
+    "Un problème est survenu": "A problem occurred",
+    "Un spectacle doit avoir au moins un tarif pour être publié.": "A show must have at least one rate to be published.",
+    "Utilisateur": "User",
+    "Utilisateur supprimé avec succès.": "User deleted successfully.",
+    "Utilisation aujourd'hui": "Usage today",
+    "VOIR LE PANIER": "VIEW CART",
+    "Valider la commande": "Validate order",
+    "Veuillez choisir un spectacle pour laisser un commentaire.": "Please choose a show to leave a comment.",
+    "Veuillez choisir une note.": "Please choose a rating.",
+    "Veuillez corriger les erreurs ci-dessous.": "Please correct the errors below.",
+    "Veuillez vérifier vos informations bancaires et réessayer.": "Please check your bank details and try again.",
+    "Visiter le site officiel": "Visit official site",
+    "Voir Tout": "View All",
+    "Voir le spectacle": "View show",
+    "Voir mes billets": "View my tickets",
+    "Voir mon panier": "View my cart",
+    "Voir sur le site": "View on site",
+    "Vos billets officiels - ThéâtrePlus": "Your official tickets - ThéâtrePlus",
+    "Votre Clé API": "Your API Key",
+    "Votre Critique": "Your Review",
+    "Votre Panier": "Your Cart",
+    "Votre Panier - ThéâtrePlus": "Your Cart - ThéâtrePlus",
+    "Votre abonnement API a été mis à jour avec succès.": "Your API subscription has been successfully updated.",
+    "Votre avis a été soumis avec succès et sera publié après validation.": "Your review has been submitted and will be published after validation.",
+    "Votre commentaire": "Your comment",
+    "Votre motivation": "Your motivation",
+    "Votre paiement a été traité avec succès.": "Your payment was processed successfully.",
+    "Votre panier est vide": "Your cart is empty",
+    "Votre panier est vide.": "Your cart is empty.",
+    "Votre quota est réinitialisé chaque jour à minuit.": "Your quota is reset every day at midnight.",
+    "Votre transaction a été refusée par l'organisme bancaire.": "Your transaction was rejected by the bank.",
+    "Vous avez choisi d'interrompre le processus de paiement.": "You chose to interrupt the payment process.",
+    "Vous n'avez pas encore activé votre accès API.": "You have not activated your API access yet.",
+    "Vous n'avez pas encore effectué de réservation.": "You haven't made any bookings yet.",
+    "Vous n'avez pas encore partagé de spectacle.": "You haven't shared any shows yet.",
+    "Vous n'avez pas encore rédigé d'article.": "You haven't written any articles yet.",
+    "Vous n'avez pas l'autorisation d'accéder à cette page!": "You are not authorized to access this page!",
+    "Vous n'êtes pas autorisé à supprimer ce message.": "You are not authorized to delete this message.",
+    "Vous ne pouvez commenter que les spectacles auxquels vous avez assisté.": "You can only comment on shows you have attended.",
+    "Vous pouvez les consulter même si vous n'avez plus accès à internet.": "You can view them even without internet access.",
+    "Vous êtes déjà critique de presse.": "You are already a press critic.",
+    "Vous êtes déjà inscrit!": "You are already registered!",
+    "Vous êtes déjà producteur.": "You are already a producer.",
+    "appels / jour": "calls / day",
+    "appels par jour.": "calls per day.",
+    "du": "from",
+    "minutes": "minutes",
+    "mois": "month",
+    "place(s) à": "seat(s) at",
+    "places": "seats",
+    "places disponibles": "available seats",
+    "sur": "on",
+    "À propos": "About",
+    "À propos de ThéâtrePlus": "About ThéâtrePlus",
+    "ÉCHEC": "FAILURE",
+    "ÉCHOUÉ": "FAILED",
+    "ÉPINGLÉ": "PINNED",
+    "Échec du Paiement": "Payment Failure",
+    "Échec du paiement - ThéâtrePlus": "Payment failed - ThéâtrePlus",
+    "Éditer": "Edit",
+    "Éditer (Admin)": "Edit (Admin)",
+    "Épingler": "Pin",
+    "Épingler en haut": "Pin to top",
+    "Épinglé": "Pinned",
+    "État": "State",
+    "Événement": "Event",
+    "Événements à venir": "Upcoming events",
+    "Êtes-vous sûr ?": "Are you sure?",
+    "Êtes-vous sûr de vouloir refuser cette candidature ?": "Are you sure you want to reject this application?",
+    "Êtes-vous sûr de vouloir refuser cette demande ?": "Are you sure you want to reject this request?",
+    "à": "at",
+    "à partir de": "from",
+}
+
+apply_batch('locale/nl/LC_MESSAGES/django.po', batch4_nl)
+apply_batch('locale/en/LC_MESSAGES/django.po', batch4_en)
+print("Batch 4 (R-Z) applied.")
