--- translate_batch2.py
@@ -1,214 +1,0 @@
-import os
-from translate_batch1 import apply_batch
-
-# BATCH 2: NL
-batch2_nl = {
-    "Dashboard Admin": "Admin Dashboard",
-    "Date": "Datum",
-    "Date & Heure": "Datum & Tijd",
-    "Date :": "Datum:",
-    "Date Soumission": "Datum van indiening",
-    "Date inconnue": "Onbekende datum",
-    "Demande en cours de traitement": "Aanvraag wordt verwerkt",
-    "Dernières places :": "Laatste plaatsen:",
-    "Description": "Beschrijving",
-    "Devenir Critique de Presse": "Persrecensent worden",
-    "Devenir Critique de Presse ?": "Persrecensent worden?",
-    "Devenir Producteur": "Producent worden",
-    "Disponible": "Beschikbaar",
-    "Disponibles": "Beschikbaar",
-    "Données complètes": "Volledige gegevens",
-    "Données limitées": "Beperkte gegevens",
-    "Durée (min)": "Duur (min)",
-    "Durée (minutes)": "Duur (minuten)",
-    "Durée :": "Duur:",
-    "Dutch": "Nederlands",
-    "Déconnexion": "Uitloggen",
-    "Découvrez les salles qui accueillent nos événements passionnants.": "Ontdek de zalen waar onze spannende evenementen plaatsvinden.",
-    "Découvrez notre mission et notre histoire": "Ontdek onze missie en onze geschiedenis",
-    "Découvrez tous nos événements passionnants.": "Ontdek al onze spannende evenementen.",
-    "Découvrir les spectacles": "Ontdek de voorstellingen",
-    "Désolé, il ne reste que %(seats)s places disponibles.": "Sorry, er zijn nog maar %(seats)s plaatsen beschikbaar.",
-    "Désolé, nous n'avons pas pu trouver les détails de ce spectacle.": "Sorry, we konden de details van deze voorstelling niet vinden.",
-    "Désépingler": "Vastzetten ongedaan maken",
-    "Détails": "Details",
-    "Détails de la candidature": "Details van de kandidatuur",
-    "Détails de la demande": "Details van de aanvraag",
-    "Détails de votre réservation": "Details van uw reservering",
-    "Détails du Spectacle": "Details van de voorstelling",
-    "Détails et informations pratiques du lieu": "Details en praktische informatie van de locatie",
-    "EN ATTENTE": "IN AFWACHTING",
-    "Email": "E-mail",
-    "En attente": "In afwachting",
-    "Endpoint": "Eindpunt",
-    "English": "Engels",
-    "Enregistrer": "Opslaan",
-    "Enregistrer les modifications": "Wijzigingen opslaan",
-    "Envoyer ma candidature": "Mijn kandidatuur verzenden",
-    "Espace Développeur": "Ontwikkelaarsruimte",
-    "Espace Producteur": "Producentenruimte",
-    "Examiner": "Beoordelen",
-    "Explorez les événements à venir et réservez vos billets.": "Ontdek de komende evenementen en reserveer uw tickets.",
-    "FAQ": "Veelgestelde vragen",
-    "Fermer": "Sluiten",
-    "Festivals": "Festivals",
-    "Fiche d'un lieu de spectacle": "Informatieblad van een locatie",
-    "Fiche d'un spectacle": "Informatieblad van een voorstelling",
-    "Fiche d'une critique": "Informatieblad van een recensie",
-    "Finalisez votre réservation et préparez-vous pour le spectacle.": "Rond uw reservering af en bereid u voor op de voorstelling.",
-    "Frais de service": "Servicekosten",
-    "Français": "Frans",
-    "French": "Frans",
-    "Gestion des Tarifs": "Tariefbeheer",
-    "Générer ma clé API Free": "Genereer mijn gratis API-sleutel",
-    "Gérer les Articles de Presse": "Persartikelen beheren",
-    "Gérez les candidatures pour devenir critique de presse.": "Beheer de kandidaturen om persrecensent te worden.",
-    "Gérez les demandes des utilisateurs pour devenir producteur.": "Beheer de aanvragen van gebruikers om producent te worden.",
-    "Gérez les nouvelles propositions des producteurs.": "Beheer de nieuwe voorstellen van de producenten.",
-    "Gérez vos critiques et analyses de spectacles.": "Beheer uw recensies en analyses van voorstellingen.",
-    "Gérez vos informations et consultez vos réservations.": "Beheer uw informatie en bekijk uw reserveringen.",
-    "Gérez vos spectacles et suivez vos performances.": "Beheer uw voorstellingen en volg uw prestaties.",
-    "Gérez votre accès API et vos abonnements d'affiliation.": "Beheer uw API-toegang en affiliate-abonnementen.",
-    "Heure": "Tijd",
-    "Historique des Paiements Stripe": "Stripe Betalingsgeschiedenis",
-    "Historique récent": "Recente geschiedenis",
-    "ID": "ID",
-    "ID Session Stripe": "Stripe Sessie ID",
-    "Il ne reste que %(count)s places pour cette séance. Ne tardez pas !": "Er zijn nog maar %(count)s plaatsen voor deze sessie. Wacht niet te lang!",
-    "Il semble que vous n'ayez pas encore choisi de spectacle.": "Het lijkt erop dat u nog geen voorstelling heeft gekozen.",
-    "Image actuelle": "Huidige afbeelding",
-    "Image du spectacle": "Afbeelding van de voorstelling",
-    "Imprimer": "Printen",
-    "Informations": "Informatie",
-    "Informations du Spectacle": "Informatie over de voorstelling",
-    "Inscription": "Registreren",
-    "L'actualité et les critiques de la scène théâtrale.": "Het nieuws en de recensies van de theaterwereld.",
-    "La culture à portée de main": "Cultuur binnen handbereik",
-    "Langue préférée": "Voorkeurstaal",
-    "Le corps de votre article. Soyez précis et constructif.": "De inhoud van uw artikel. Wees nauwkeurig en constructief.",
-    "Le nombre de tickets ne peut pas dépasser la capacité de la salle.": "Het aantal tickets mag de capaciteit van de zaal niet overschrijden.",
-    "Lien Ticketmaster": "Ticketmaster-link",
-    "Lieu": "Locatie",
-    "Lieu :": "Locatie:",
-    "Lieu créé avec succès.": "Locatie succesvol aangemaakt.",
-    "Lieu inconnu": "Onbekende locatie",
-    "Lieu inexistant": "Locatie bestaat niet",
-    "Lieu modifié avec succès.": "Locatie succesvol gewijzigd.",
-    "Lieu non défini": "Locatie niet gedefinieerd",
-    "Lieu supprimé avec succès.": "Locatie succesvol verwijderd.",
-    "Lieux": "Locaties",
-    "Lire la suite": "Lees meer",
-    "Liste de mes articles": "Lijst van mijn artikelen",
-    "Liste des critiques": "Lijst van recensies",
-    "Liste des lieux de spectacle": "Lijst van locaties",
-    "Liste des spectacles": "Lijst van voorstellingen",
-    "Ma Consommation": "Mijn Verbruik",
-}
-
-# BATCH 2: EN
-batch2_en = {
-    "Dashboard Admin": "Admin Dashboard",
-    "Date": "Date",
-    "Date & Heure": "Date & Time",
-    "Date :": "Date:",
-    "Date Soumission": "Submission Date",
-    "Date inconnue": "Unknown date",
-    "Demande en cours de traitement": "Request being processed",
-    "Dernières places :": "Last seats:",
-    "Description": "Description",
-    "Devenir Critique de Presse": "Become a Press Critic",
-    "Devenir Critique de Presse ?": "Become a Press Critic?",
-    "Devenir Producteur": "Become a Producer",
-    "Disponible": "Available",
-    "Disponibles": "Available",
-    "Données complètes": "Full data",
-    "Données limitées": "Limited data",
-    "Durée (min)": "Duration (min)",
-    "Durée (minutes)": "Duration (minutes)",
-    "Durée :": "Duration:",
-    "Dutch": "Dutch",
-    "Déconnexion": "Logout",
-    "Découvrez les salles qui accueillent nos événements passionnants.": "Discover the venues hosting our exciting events.",
-    "Découvrez notre mission et notre histoire": "Discover our mission and history",
-    "Découvrez tous nos événements passionnants.": "Discover all our exciting events.",
-    "Découvrir les spectacles": "Discover shows",
-    "Désolé, il ne reste que %(seats)s places disponibles.": "Sorry, only %(seats)s seats remaining.",
-    "Désolé, nous n'avons pas pu trouver les détails de ce spectacle.": "Sorry, we could not find details for this show.",
-    "Désépingler": "Unpin",
-    "Détails": "Details",
-    "Détails de la candidature": "Application details",
-    "Détails de la demande": "Request details",
-    "Détails de votre réservation": "Your reservation details",
-    "Détails du Spectacle": "Show Details",
-    "Détails et informations pratiques du lieu": "Venue details and practical information",
-    "EN ATTENTE": "PENDING",
-    "Email": "Email",
-    "En attente": "Pending",
-    "Endpoint": "Endpoint",
-    "English": "English",
-    "Enregistrer": "Save",
-    "Enregistrer les modifications": "Save changes",
-    "Envoyer ma candidature": "Send my application",
-    "Espace Développeur": "Developer Space",
-    "Espace Producteur": "Producer Space",
-    "Examiner": "Review",
-    "Explorez les événements à venir et réservez vos billets.": "Explore upcoming events and book your tickets.",
-    "FAQ": "FAQ",
-    "Fermer": "Close",
-    "Festivals": "Festivals",
-    "Fiche d'un lieu de spectacle": "Venue information",
-    "Fiche d'un spectacle": "Show information",
-    "Fiche d'une critique": "Review information",
-    "Finalisez votre réservation et préparez-vous pour le spectacle.": "Finalize your booking and get ready for the show.",
-    "Frais de service": "Service fee",
-    "Français": "French",
-    "French": "French",
-    "Gestion des Tarifs": "Price Management",
-    "Générer ma clé API Free": "Generate my Free API Key",
-    "Gérer les Articles de Presse": "Manage Press Articles",
-    "Gérez les candidatures pour devenir critique de presse.": "Manage applications to become a press critic.",
-    "Gérez les demandes des utilisateurs pour devenir producteur.": "Manage user requests to become a producer.",
-    "Gérez les nouvelles propositions des producteurs.": "Manage new producer proposals.",
-    "Gérez vos critiques et analyses de spectacles.": "Manage your show reviews and analyses.",
-    "Gérez vos informations et consultez vos réservations.": "Manage your info and view your bookings.",
-    "Gérez vos spectacles et suivez vos performances.": "Manage your shows and track performance.",
-    "Gérez votre accès API et vos abonnements d'affiliation.": "Manage your API access and affiliate plans.",
-    "Heure": "Time",
-    "Historique des Paiements Stripe": "Stripe Payment History",
-    "Historique récent": "Recent history",
-    "ID": "ID",
-    "ID Session Stripe": "Stripe Session ID",
-    "Il ne reste que %(count)s places pour cette séance. Ne tardez pas !": "Only %(count)s seats left for this session. Don't wait!",
-    "Il semble que vous n'ayez pas encore choisi de spectacle.": "It seems you haven't chosen a show yet.",
-    "Image actuelle": "Current image",
-    "Image du spectacle": "Show image",
-    "Imprimer": "Print",
-    "Informations": "Information",
-    "Informations du Spectacle": "Show Information",
-    "Inscription": "Sign Up",
-    "L'actualité et les critiques de la scène théâtrale.": "Latest news and reviews from the theater scene.",
-    "La culture à portée de main": "Culture at your fingertips",
-    "Langue préférée": "Preferred language",
-    "Le corps de votre article. Soyez précis et constructif.": "Your article body. Be precise and constructive.",
-    "Le nombre de tickets ne peut pas dépasser la capacité de la salle.": "Ticket count cannot exceed room capacity.",
-    "Lien Ticketmaster": "Ticketmaster Link",
-    "Lieu": "Venue",
-    "Lieu :": "Venue:",
-    "Lieu créé avec succès.": "Venue created successfully.",
-    "Lieu inconnu": "Unknown venue",
-    "Lieu inexistant": "Venue does not exist",
-    "Lieu modifié avec succès.": "Venue updated successfully.",
-    "Lieu non défini": "Venue not defined",
-    "Lieu supprimé avec succès.": "Venue deleted successfully.",
-    "Lieux": "Venues",
-    "Lire la suite": "Read more",
-    "Liste de mes articles": "List of my articles",
-    "Liste des critiques": "List of reviews",
-    "Liste des lieux de spectacle": "List of venues",
-    "Liste des spectacles": "List of shows",
-    "Ma Consommation": "My Usage",
-}
-
-apply_batch('locale/nl/LC_MESSAGES/django.po', batch2_nl)
-apply_batch('locale/en/LC_MESSAGES/django.po', batch2_en)
-print("Batch 2 (E-M) applied.")
